Clean .env dependencies
