+++ .env.example
@@ -1,18 +1,19 @@
-DEPLOYER_PRIV_KEY=     # required
-USER_PRIV_KEY=         # optional
+# tests
+DEPLOYER_PRIV_KEY=        # required for tests
+USER_PRIV_KEY=            # optional for tests
+TEST_ARBITRUM_L1_RPC_URL= # required for tests
+TEST_ARBITRUM_L2_RPC_URL='https://rinkeby.arbitrum.io/rpc'
+TEST_OPTIMISM_L1_RPC_URL='http://127.0.0.1:9545'
+TEST_OPTIMISM_L2_RPC_URL='http://127.0.0.1:8545'
 
-ARBITRUM_L1_RPC_URL=
-ARBITRUM_L2_RPC_URL='https://rinkeby.arbitrum.io/rpc'
+# deployment
+ORACLE_PRIV_KEY=                    # required to test deployment
+RECEIVER_PRIV_KEY=                  # required to test deployment
+KOVAN_OPTIMISM_DEPLOYER_PRIV_KEY=   # required for deployment
+KOVAN_OPTIMISM_L1_RPC=              # required for deployment
+KOVAN_OPTIMISM_L2_RPC=              # required for deployment
+RINKEBY_ARBITRUM_DEPLOYER_PRIV_KEY= # required for deployment
+RINKEBY_ARBITRUM_L1_RPC=            # required for deployment
+RINKEBY_ARBITRUM_L2_RPC=            # required for deployment
 
-OPTIMISM_L1_RPC_URL='http://127.0.0.1:9545'
-OPTIMISM_L2_RPC_URL='http://127.0.0.1:8545'
-
-KOVAN_OPTIMISM_DEPLOYER_PRIV_KEY=
-KOVAN_OPTIMISM_L1_RPC=
-KOVAN_OPTIMISM_L2_RPC=
-
-RINKEBY_ARBITRUM_DEPLOYER_PRIV_KEY=
-RINKEBY_ARBITRUM_L1_RPC=
-RINKEBY_ARBITRUM_L2_RPC=
-
-ETHERSCAN_KEY=
+ETHERSCAN_KEY= # optional
