+++ test/optimism/setup.ts
@@ -1,8 +1,11 @@
 import { getMainnetSdk } from '@dethcrypto/eth-sdk-client'
 import { JsonRpcProvider } from '@ethersproject/providers'
+import { getOptionalEnv, getRequiredEnv } from '@makerdao/hardhat-utils'
+import { ethers, Wallet } from 'ethers'
 
 import { L1AddWormholeOptimismSpell__factory, L2AddWormholeDomainSpell__factory } from '../../typechain'
 import { deployUsingFactory, forwardTime, getContractFactory, mintEther, toEthersBigNumber } from '../helpers'
+import { RetryProvider } from '../helpers/RetryProvider'
 import {
   deployWormhole,
   DomainSetupOpts,
@@ -25,11 +28,6 @@ import {
 const TTL = OPTIMISTIC_ROLLUP_FLUSH_FINALIZATION_TIME
 
 export async function setupOptimismTests({
-  l1Signer,
-  l2Signer,
-  l1User,
-  l1Provider,
-  l2Provider,
   l2DaiAmount,
   domain,
   masterDomain,
@@ -37,6 +35,26 @@ export async function setupOptimismTests({
   fee,
   line,
 }: DomainSetupOpts): Promise<DomainSetupResult> {
+  const l1Rpc = getRequiredEnv(`TEST_OPTIMISM_L1_RPC_URL`)
+  const l2Rpc = getRequiredEnv(`TEST_OPTIMISM_L2_RPC_URL`)
+
+  const pkey = getRequiredEnv('DEPLOYER_PRIV_KEY')
+  const pkey2 = getOptionalEnv('USER_PRIV_KEY')
+
+  const l1Provider = new ethers.providers.JsonRpcProvider(l1Rpc)
+  const l2Provider = new RetryProvider(5, l2Rpc)
+  const l1StartingBlock = await l1Provider.getBlockNumber()
+  const l2StartingBlock = await l2Provider.getBlockNumber()
+  console.log('Current L1 block: ', l1StartingBlock)
+  console.log('Current L2 block: ', l2StartingBlock)
+
+  const l1Signer = new ethers.Wallet(pkey, l1Provider)
+  const l2Signer = new ethers.Wallet(pkey, l2Provider)
+  const l1User = pkey2 ? new ethers.Wallet(pkey2, l1Provider) : Wallet.createRandom().connect(l1Provider)
+  const l2User = l1User.connect(l2Provider)
+  console.log('l1Signer:', l1Signer.address)
+  console.log('l1User:', l1User.address)
+
   const l1Sdk = getMainnetSdk(l1Signer)
   const makerSdk = l1Sdk.maker
   const optimismRollupSdk = await getDynamicOptimismRollupSdk(l1Signer, l2Signer)
@@ -113,6 +131,14 @@ export async function setupOptimismTests({
   )
   console.log('Optimism setup complete.')
   return {
+    l1Signer,
+    l2Signer,
+    l1User,
+    l2User,
+    l1Provider,
+    l2Provider,
+    l1StartingBlock,
+    l2StartingBlock,
     makerSdk,
     relayTxToL1,
     relayTxToL2,
