+++ test/wormhole/setup.ts
@@ -1,12 +1,10 @@
 import { MainnetSdk } from '@dethcrypto/eth-sdk-client'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { randomBytes } from '@ethersproject/random'
-import { getOptionalEnv, getRequiredEnv } from '@makerdao/hardhat-utils'
 import { BigNumber, BigNumberish, Contract, Wallet } from 'ethers'
 import { ethers } from 'hardhat'
 
 import { BasicRelay, WormholeConstantFee, WormholeJoin, WormholeOracleAuth, WormholeRouter } from '../../typechain'
-import { RetryProvider } from '../helpers/RetryProvider'
 import { BaseBridgeSdk, DaiLike, L1EscrowLike, WormholeBridgeSdk } from '.'
 import { performSanityChecks } from './checks'
 import { RelayTxToL1Function, RelayTxToL2Function } from './messages'
@@ -19,11 +17,6 @@ const masterDomain = bytes32('L1')
 export type MakerSdk = MainnetSdk['maker']
 
 export interface DomainSetupOpts {
-  l1Signer: Wallet
-  l2Signer: Wallet
-  l1User: Wallet
-  l1Provider: JsonRpcProvider
-  l2Provider: JsonRpcProvider
   l2DaiAmount: BigNumberish
   domain: string
   masterDomain: string
@@ -42,6 +35,14 @@ export interface DomainSetupResult {
   ttl: number
   forwardTimeToAfterFinalization: ForwardTimeFunction
   addWormholeDomainSpell: Contract
+  l1Signer: Wallet
+  l2Signer: Wallet
+  l1Provider: JsonRpcProvider
+  l2Provider: JsonRpcProvider
+  l1User: Wallet
+  l2User: Wallet
+  l1StartingBlock: number
+  l2StartingBlock: number
 }
 
 export type ForwardTimeFunction = (l1Provider: JsonRpcProvider) => Promise<void>
@@ -85,28 +86,15 @@ export async function setupTest({
   oracleAddresses,
   setupDomain,
 }: SetupTestOpts): Promise<SetupTestResult> {
-  const pkey = getRequiredEnv('DEPLOYER_PRIV_KEY')
-  const pkey2 = getOptionalEnv('USER_PRIV_KEY')
-  const l1Rpc = getRequiredEnv(`${ethers.utils.parseBytes32String(domain).split('-')[0].toUpperCase()}_L1_RPC_URL`)
-  const l2Rpc = getRequiredEnv(`${ethers.utils.parseBytes32String(domain).split('-')[0].toUpperCase()}_L2_RPC_URL`)
-
-  const l1Provider = new ethers.providers.JsonRpcProvider(l1Rpc)
-  const l2Provider = new RetryProvider(5, l2Rpc)
-  const l1StartingBlock = await l1Provider.getBlockNumber()
-  const l2StartingBlock = await l2Provider.getBlockNumber()
-  console.log('Current L1 block: ', l1StartingBlock)
-  console.log('Current L2 block: ', l2StartingBlock)
-
-  const l1Signer = new ethers.Wallet(pkey, l1Provider)
-  const l2Signer = new ethers.Wallet(pkey, l2Provider)
-  const l1User = pkey2 ? new ethers.Wallet(pkey2, l1Provider) : Wallet.createRandom().connect(l1Provider)
-  const l2User = l1User.connect(l2Provider)
-  console.log('l1Signer:', l1Signer.address)
-  console.log('l1User:', l1User.address)
-
   const ilk: string = bytes32('WH_' + Buffer.from(randomBytes(14)).toString('hex'))
 
   const {
+    l1Signer,
+    l1User,
+    l2User,
+    l1Provider,
+    l1StartingBlock,
+    l2StartingBlock,
     makerSdk,
     relayTxToL1,
     relayTxToL2,
@@ -117,11 +105,6 @@ export async function setupTest({
     forwardTimeToAfterFinalization,
     addWormholeDomainSpell,
   } = await setupDomain({
-    l1Signer,
-    l2Signer,
-    l1User,
-    l1Provider,
-    l2Provider,
     l2DaiAmount,
     domain,
     masterDomain,
